Make Doorkeeper routes mapping reusable in extensions
